feat: ImageBox 컴포넌트 구현
- 등수에 따라서 이벤트 이미지에 뱃지를 붙여 보여주는
  ImageBox 컴포넌트 구현
