+++ src/components/pages/Event/PostCard.tsx
@@ -1,5 +1,6 @@
 import { IUserItem } from "@/components/common/UserItem";
 import EventPostFooter from "./PostFooter";
+import ImageBox from "./ImageBox";
 
 type TRank = "1st" | "2nd" | "3rd" | "none";
 
@@ -11,19 +12,6 @@ interface IEventPostCard {
   rank?: TRank;
 }
 
-function renderBadge(rank: TRank) {
-  switch (rank) {
-    case "1st":
-      return <img src={"/icon/badge/first-badge.PNG"} alt="badge" />;
-    case "2nd":
-      return <img src={"/icon/badge/second-badge.PNG"} alt="badge" />;
-    case "3rd":
-      return <img src={"/icon/badge/third-badge.PNG"} alt="badge" />;
-    default:
-      return null;
-  }
-}
-
 export default function EventPostCard({
   postId,
   postImageUrl,
@@ -33,14 +21,7 @@ export default function EventPostCard({
 }: IEventPostCard) {
   return (
     <div className="relative w-[100px] flex flex-col gap-1">
-      <div className="absolute top-1 left-1 w-[20px]">{renderBadge(rank)}</div>
-      <div className="w-full h-[100px] rounded-lg overflow-hidden border border-muted-foreground">
-        <img
-          src={postImageUrl}
-          alt="postImage"
-          className="w-full h-full object-cover"
-        />
-      </div>
+      <ImageBox src={postImageUrl} rank={rank} />
       <EventPostFooter
         postId={postId}
         userInfo={userInfo}
